Changed the RangeValidatedType Property from Int32 to Decimal as this is the smallest possible change in the next series of fixes around the [Range] attribute on properties and fields of type Decimal.

+++ Src/AutoFixtureUnitTest/DataAnnotations/RangeValidatedType.cs
@@ -8,6 +8,6 @@ namespace Ploeh.AutoFixtureUnitTest.DataAnnotations
         public const int Maximum = 20;
 
         [Range(Minimum, Maximum)]
-        public int Property { get; set; }
+        public decimal Property { get; set; }
     }
 }
